Added environment modes and .env

+++ index.js
@@ -4,13 +4,18 @@ const router = require('./routers/');
 const passport = require('./core/auth');
 const sequelize = require('./core/db');
 
+const isProduction = process.env.NODE_ENV === 'production';
+
 const app = new Koa();
 
-app.use(logger());
+if (!isProduction) {
+  app.use(logger());
+}
+
 app.use(router());
 app.use(passport.initialize());
 
-app.listen(3000, async () => {
-  await sequelize.sync({ force: false });
-  console.log('Server running on https://localhost:3000')
+app.listen(process.env.PORT, async () => {
+  if (!isProduction) await sequelize.sync({ force: false });
+  console.log(`Server running on https://localhost:${process.env.PORT}`)
 });
